+++ webpack/config/modules.js
